// cadriciel/client/src/app/enums/tools.ts
export enum Tools {
    Pencil = 'pencil',
    Brush = 'brush',
    Pen = 'pen',
    Feather = 'feather',
    Aerosol = 'aerosol',
    Rectangle = 'rectangle',
    Ellipse = 'ellipse',
    Polygon = 'polygon',
    Selection = 'selection',
    ColorApplicator = 'colorApplicator',
    ColorBucket = 'colorBucket',
    Eraser = 'eraser',
    Text = 'text',
    Pipette = 'pipette',
    Line = 'line',
    Stamp = 'stamp',
    Grid = 'grid',
    None = 'none',
}